attempting deduplication
