+++ src/Schema/Attribute/Group.php
@@ -2,57 +2,10 @@
 
 namespace GoetasWebservices\XML\XSDReader\Schema\Attribute;
 
+use GoetasWebservices\XML\XSDReader\Schema\AbstractNamedGroupItem;
 use GoetasWebservices\XML\XSDReader\Schema\Schema;
 
-class Group implements AttributeItem, AttributeContainer
+class Group extends AbstractNamedGroupItem implements AttributeItem, AttributeContainer
 {
-    use AttributeItemTrait;
     use AttributeContainerTrait;
-
-    /**
-     * @var Schema
-     */
-    protected $schema;
-
-    /**
-     * @var string|null
-     */
-    protected $doc;
-
-    /**
-     * @param string $name
-     */
-    public function __construct(Schema $schema, $name)
-    {
-        $this->schema = $schema;
-        $this->name = $name;
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getDoc()
-    {
-        return $this->doc;
-    }
-
-    /**
-     * @param string $doc
-     *
-     * @return $this
-     */
-    public function setDoc($doc)
-    {
-        $this->doc = $doc;
-
-        return $this;
-    }
-
-    /**
-     * @return Schema
-     */
-    public function getSchema()
-    {
-        return $this->schema;
-    }
 }
